+++ .idea/compiler.xml
@@ -7,6 +7,7 @@
         <sourceOutputDir name="target/generated-sources/annotations" />
         <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
         <outputRelativeToContentRoot value="true" />
+        <module name="Labygame_project" />
         <module name="labygame" />
       </profile>
     </annotationProcessing>
