+++ src/main/java/com/labygame/personnage/Monster.java
@@ -1,4 +1,4 @@
-package com.labygame.demo.personnage;
+package com.labygame.personnage;
 
 import java.util.Random;
