+++ src/main/java/com/labygame/trayEnvironnement/gameTrayPiece.java
@@ -1,4 +1,4 @@
-package com.labygame.demo.trayEnvironnement;
+package com.labygame.trayEnvironnement;
 
 public class gameTrayPiece {
     //a remplir
