+++ src/main/java/module-info.java
@@ -4,8 +4,7 @@ module com.labygame.demo {
     requires static lombok;
     requires org.jetbrains.annotations;
     requires com.google.gson;
-    requires org.junit.jupiter.api;
 
-    opens com.labygame.demo to javafx.fxml;
-    exports com.labygame.demo;
+    opens com.labygame to javafx.fxml;
+    exports com.labygame;
 }
